Toolbar Placement working for tabbed and nav pages
