+++ Xamarin.Forms.Platform.WinRT.Phone/FormsPivot.cs
@@ -2,6 +2,7 @@
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media;
+using Xamarin.Forms.PlatformConfiguration.WindowsSpecific;
 
 #if WINDOWS_UWP
 
@@ -12,7 +13,7 @@ namespace Xamarin.Forms.Platform.WinRT
 {
 	public class FormsPivot : Pivot, IToolbarProvider
 	{
-		public static readonly DependencyProperty ToolbarVisibilityProperty = DependencyProperty.Register(nameof(ToolbarVisibility), typeof(Visibility), typeof(FormsPivot),
+		public static readonly DependencyProperty TitleVisibilityProperty = DependencyProperty.Register(nameof(TitleVisibility), typeof(Visibility), typeof(FormsPivot),
 			new PropertyMetadata(Visibility.Collapsed));
 
 		public static readonly DependencyProperty ToolbarForegroundProperty = DependencyProperty.Register(nameof(ToolbarForeground), typeof(Brush), typeof(FormsPivot), new PropertyMetadata(default(Brush)));
@@ -22,8 +23,10 @@ namespace Xamarin.Forms.Platform.WinRT
 		CommandBar _commandBar;
 
 		TaskCompletionSource<CommandBar> _commandBarTcs;
+	    ToolbarPlacement _toolbarPlacement;
+	    ContentControl _titleBar;
 
-		public Brush ToolbarBackground
+	    public Brush ToolbarBackground
 		{
 			get { return (Brush)GetValue(ToolbarBackgroundProperty); }
 			set { SetValue(ToolbarBackgroundProperty, value); }
@@ -35,11 +38,21 @@ namespace Xamarin.Forms.Platform.WinRT
 			set { SetValue(ToolbarForegroundProperty, value); }
 		}
 
-		public Visibility ToolbarVisibility
+		public Visibility TitleVisibility
 		{
-			get { return (Visibility)GetValue(ToolbarVisibilityProperty); }
-			set { SetValue(ToolbarVisibilityProperty, value); }
+			get { return (Visibility)GetValue(TitleVisibilityProperty); }
+			set { SetValue(TitleVisibilityProperty, value); }
 		}
+
+        public ToolbarPlacement ToolbarPlacement
+	    {
+	        get { return _toolbarPlacement; }
+	        set
+	        {
+	            _toolbarPlacement = value;
+	            UpdateToolbarPlacement();
+	        }
+	    }
 
 		Task<CommandBar> IToolbarProvider.GetCommandBarAsync()
 		{
@@ -55,10 +68,53 @@ namespace Xamarin.Forms.Platform.WinRT
 		{
 			base.OnApplyTemplate();
 			_commandBar = GetTemplateChild("CommandBar") as CommandBar;
+            _titleBar = GetTemplateChild("TitleBar") as ContentControl;
 			TaskCompletionSource<CommandBar> tcs = _commandBarTcs;
-			if (tcs != null)
+		    tcs?.SetResult(_commandBar);
+		}
+
+        void UpdateToolbarPlacement()
+		{
+			if (_commandBar == null)
 			{
-				tcs.SetResult(_commandBar);
+				return;
+			}
+
+			switch (ToolbarPlacement)
+			{
+				case ToolbarPlacement.Top:
+					Windows.UI.Xaml.Controls.Grid.SetRow(_commandBar, 0);
+					break;
+				case ToolbarPlacement.Bottom:
+					Windows.UI.Xaml.Controls.Grid.SetRow(_commandBar, 2);
+					break;
+				case ToolbarPlacement.Default:
+				default:
+					Windows.UI.Xaml.Controls.Grid.SetRow(_commandBar, Device.Idiom == TargetIdiom.Phone ? 2 : 0);
+					break;
+			}
+
+			AdjustCommandBarForTitle();
+		}
+
+		void AdjustCommandBarForTitle()
+		{
+		    if (_commandBar == null || _titleBar == null)
+		    {
+		        return;
+		    }
+
+		    if (Windows.UI.Xaml.Controls.Grid.GetRow(_commandBar) == 0)
+			{
+				Windows.UI.Xaml.Controls.Grid.SetColumn(_commandBar, 1);
+				Windows.UI.Xaml.Controls.Grid.SetColumnSpan(_commandBar, 1);
+				Windows.UI.Xaml.Controls.Grid.SetColumnSpan(_titleBar, 1);
+			}
+			else
+			{
+				Windows.UI.Xaml.Controls.Grid.SetColumn(_commandBar, 0);
+				Windows.UI.Xaml.Controls.Grid.SetColumnSpan(_commandBar, 2);
+				Windows.UI.Xaml.Controls.Grid.SetColumnSpan(_titleBar, 2);
 			}
 		}
 	}
